refactored player class to function under an abstract class
